Code commenting, moving things into files, and a test
